// src/common/enums/analysis-server-type.enum.ts
export enum AnalysisServerType {
  HAIR = 'hair',
  SKIN = 'skin',
  CMA_HAIR = 'cmahair',
  CMA_SKIN = 'cmaskin',
  COLLECTOR = 'collector',
}

export enum AnalysisDatabaseKey {
  ANALYSIS_CNDP_HAIR = 'analysis_cndp_hair',
  ANALYSIS_CNDP_SKIN = 'analysis_cndp_skin',
  ANALYSIS_CMA_HAIR = 'analysis_cma_hair',
  ANALYSIS_CMA_SKIN = 'analysis_cma_skin',
  ANALYSIS_COLLECTOR = 'analysis_collector',
}